add class for build Graph relationship between "in circles" and "have in circles".

+++ Core/IStater.cs
@@ -5,12 +5,12 @@ using System.Text;
 
 namespace Core {
 	public interface IStater {
-		int Position {
+		long Position {
 			get;
 			set;
 		}
 
-		int MaxPosition {
+		long MaxPosition {
 			get;
 			set;
 		}
@@ -20,12 +20,12 @@ namespace Core {
 			set;
 		}
 
-		void Inc(int value = 1);
+		void Inc(long value = 1);
 
-		void MovePosition(int newPosition);
+		void MovePosition(long newPosition);
 
-		void SetState(string description, int position = 0, bool incPosition = false);
+		void SetState(string description, long position = 0, bool incPosition = false);
 
-		void InitState(string description = default(string), int maxPosition = 100, int startPosition = 0);
+		void InitState(string description = default(string), long maxPosition = 100, long startPosition = 0);
 	}
 }
